Run database seed operations as www-data
Previously the commands were run as root causing the database to be
unwritable by the Rails app instance. This will cause them to be owned
by the www-data user and group.

Note: This change depends on docker/docker#12025, in Docker release 1.7.0.
Pre-1.7.0 the default user was the run-as-user for the container instead
of root.